feat: implement responsive navigation with separate desktop and mobile components
